refactor(api): split handlers_survey.go (1720 → 7 files) by concern
Third worst-first file-size refactor. Function bodies relocated unchanged; a
few drive-by lint fixes were forced by --new-from-rev on the moved lines.

  handlers_survey.go              408   ctx helpers + survey CRUD + state machine
  handlers_survey_samples.go      144   sample type detection + add-sample
  handlers_survey_floorplan.go    271   floor plan + settings + AirMapper import
  handlers_survey_analysis.go     203   dead-zones + heatmap
  handlers_survey_floors.go       362   floor list/add/get/update/delete + dispatch
  handlers_survey_floors_data.go  287   active floor + floor plan + floor sample
  handlers_survey_report.go       132   PDF report generation

All under the 600-line warn budget.

Drive-by cleanups forced by --new-from-rev on relocated lines:

- Replaced map[string]string{"status":"deleted"} / {"status":"sample added"}
  ack responses with the typed statusResponse plus two package-level
  constants (statusDeleted, statusSampleAdded) — silences goconst on the
  newly-counted occurrences without changing any wire format.
- Converted ~49 logger.Error/Warn/Info/Debug calls to their *Context
  variants threading r.Context(); processSampleData and friends now take
  ctx as their first parameter, callers pass r.Context().
- Suppressed two false-positive gosec findings inline with justification:
  G705 (XSS via Write) on PNG heatmap and PDF report writes (Content-Type
  already pinned), and G120 (unbounded ParseMultipartForm) where the body
  is already bounded above via http.MaxBytesReader.

Build, vet, lint (--new-from-rev=HEAD~1: 0 issues), and api tests pass.
